Thêm chức năng quản lý điện nước - dịch vụ

+++ src/main/java/com/example/DACK_QuanLyNhaTro/repository/PhongRepository.java
@@ -3,5 +3,5 @@ package com.example.DACK_QuanLyNhaTro.repository;
 import com.example.DACK_QuanLyNhaTro.entity.Phong;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhongRepository extends JpaRepository<Phong, Integer> {
+public interface PhongRepository extends JpaRepository<Phong, Long> {
 }
